+++ src/new-components/dropdown/dropdown.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import { Dropdown, DropdownOption } from './dropdown';
-import { Icon } from '../../new-components/icon/icon';
+import { Icon } from '../icon/icon';
 
 <Meta
   title="Forms/Dropdown"
