+++ src/new-components/file-dropper/file-list-item.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { theme } from '../layout/theme';
-import { TextStyles } from '../typography/typography-v2';
+import { theme } from '../../components/layout/theme';
+import { TextStyles } from '../../components/typography/typography-v2';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
 import { Dropdown, DropdownOption } from '../dropdown/dropdown';
